get_in_t.c: removed unused variable. rd_cred.c: removed INTERFACE keyword. ser_auth.c: passing int32 where a size_t is wanted. ser_ctx.c: 16/32 bit int size mismatch.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@6826 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/ser_ctx.c
@@ -283,7 +283,7 @@ krb5_context_internalize(kcontext, argp, buffer, lenremain)
 		/* Get the number of ktypes */
 		if (!(kret = krb5_ser_unpack_int32(&ibuf, &bp, &remain))) {
 		    /* Reduce it to a count */
-		    context->ktype_count = ibuf;
+		    context->ktype_count = (int) ibuf;
 		    if ((context->ktypes = (krb5_enctype *)
 			 malloc(sizeof(krb5_enctype) *
 				(context->ktype_count+1)))) {
